refactor: move error handling functions in error module
